tips&forum page
created tips page and worked on it. @Stephen Wildrick I need your help with a few changes I need to make. 6pm??? :smile:
created time management page, didnt work on it yet, but planning on tomorrow work and finish.
made a underline as a hover state of the nav bars
connected the navbar buttons to the pages
connected our pages to the font awesome icons and used 2 in the  tips page - heart and comment
@Varshini and @Chandy Olesen please look at the pages since I dont have access to the prototype I don't know much what to add as far as copy.